+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Fixed value not being updated in some case in MRU text combos
   - Cobol open file dialog better error and extensions handling
   - Do not persist page size when opening files
+  - Fixed charset autodetection
 - General:
   - Cobol open file dialog supports dropping files
   - Cobol view subheaders context menus and size tooltip